requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
psycopg2-binary>=2.9.9
pymongo>=4.6.0
boto3>=1.34.0
pydantic>=2.7.0
pandas>=2.2.2
tqdm>=4.66.0
faker>=24.0.0
python-dotenv>=1.0.1
